Update amspaper.tex
Matt, I am not sure that I am using GITHUB correctly. I am in a place where I have to write up something about this for our annual report. Alexia, started a white paper, anticipating your results, and I can just take it that far.  I was reading this, and 1) The beginning is excellent.  2) I do not see anything beyond discipline and focus to finish this.  It seems statistically and scientifically strong enough for review.  If there is something that yo know is weak, then tell me.  If it is as strong as I think, then I think that you would feel good to throw into into the vagaries of peer review.  3) Your revision is not far enough along that I can carry it farther, so you would have to carry it somewhat farther.